requirements.txt
PyQt6>=6.4.0
requests>=2.28.1
ffmpeg-python>=0.2.0
